requirements.txt
watchdog>=4.0.0
Pillow>=10.0.0
zwoasi>=0.2.0
numpy>=1.24.0
opencv-python>=4.8.0
ttkbootstrap>=1.10.1
requests>=2.31.0
